Move PortalDesktop to portals/ directory
